Better organize FE code
